+++ hibernate-core/src/test/java/org/hibernate/test/annotations/id/entities/TwinkleToes.java
@@ -1,6 +1,7 @@
 //$Id$
 package org.hibernate.test.annotations.id.entities;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -26,7 +27,7 @@ public class TwinkleToes implements Serializable {
 	@GeneratedValue(strategy = GenerationType.IDENTITY, generator = "java5_uuid")
 	@GenericGenerator(name = "java5_uuid", strategy = "org.hibernate.test.annotations.id.UUIDGenerator")
 	@Column(name = "id", precision = 128, scale = 0)
-	private BigInteger id;
+	private BigDecimal id;
 
 	@ManyToOne
 	@JoinColumn(name = "bunny_id")
@@ -36,7 +37,7 @@ public class TwinkleToes implements Serializable {
 		this.bunny = bunny;
 	}
 
-	public BigInteger getId() {
+	public BigDecimal getId() {
 		return id;
 	}
 }
