+++ hibernate-core/src/test/java/org/hibernate/test/annotations/id/sequences/entities/Bunny.java
@@ -1,7 +1,7 @@
 //$Id: Bunny.java 14761 2008-06-11 13:51:06Z hardy.ferentschik $
 package org.hibernate.test.annotations.id.sequences.entities;
 import java.io.Serializable;
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.Set;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
@@ -26,7 +26,7 @@ public class Bunny implements Serializable {
 	@GeneratedValue(strategy = GenerationType.IDENTITY, generator = "java5_uuid")
 	@GenericGenerator(name = "java5_uuid", strategy = "org.hibernate.test.annotations.id.UUIDGenerator")
 	@Column(name = "id", precision = 128, scale = 0)
-	private BigInteger id;
+	private BigDecimal id;
 
 	@OneToMany(mappedBy = "bunny", cascade = { CascadeType.PERSIST })
 	Set<PointyTooth> teeth;
@@ -38,7 +38,7 @@ public class Bunny implements Serializable {
 		this.teeth = teeth;
 	}
 
-	public BigInteger getId() {
+	public BigDecimal getId() {
 		return id;
 	}
 }
